Convert validatecontails to gomega
