+++ go.mod
@@ -6,10 +6,11 @@ require (
 	github.com/aelsabbahy/go-ps v0.0.0-20170721000941-443386855ca1
 	github.com/blang/semver v3.5.1+incompatible
 	github.com/cheekybits/genny v0.0.0-20160824153601-e8e29e67948b
+	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/docker/docker v0.0.0-20161109014415-383a2f046b16
 	github.com/fatih/color v0.0.0-20161025120501-bf82308e8c85
 	github.com/golang/protobuf v1.3.2 // indirect
-	github.com/jaypipes/ghw v0.6.0
+	github.com/kr/pretty v0.2.0 // indirect
 	github.com/mattn/go-colorable v0.0.9 // indirect
 	github.com/mattn/go-isatty v0.0.4 // indirect
 	github.com/miekg/dns v0.0.0-20161018060808-58f52c57ce9d
